requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
beautifulsoup4==4.12.2
billiard==4.2.0
bootstrap4==0.1.0
celery==5.3.6
channels==4.0.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
Django==4.2.7
django-bootstrap4==23.2
django-cors-headers==4.3.1
django-filter==2.4.0
django-rest-framework==0.1.0
django-seed==0.3.1
django-widget-tweaks==1.4.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
et-xmlfile==1.1.0
Faker==22.5.0
fonttools==4.47.0
greenlet==3.0.1
kiwisolver==1.4.5
kombu==5.3.4
Markdown==2.6.11
marshmallow==1.2.6
matplotlib==3.8.2
mysql-connector-python==8.2.0
mysqlclient==2.2.0
numpy==1.26.2
openpyxl==3.1.2
packaging==23.2
pandas==2.1.3
Pillow==8.3.2
prompt-toolkit==3.0.41
protobuf==4.21.12
PyJWT==2.8.0
PyMySQL==1.1.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
redis==5.0.1
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.23
sqlparse==0.4.2
toposort==1.10
typing_extensions==4.8.0
tzdata==2023.3
vine==5.1.0
wcwidth==0.2.12
xlrd==2.0.1


# para instalar las dependencias 
# pip install -r requirements.txt 
# para actualizar 
# pip freeze > requirements.txt